# backend/api/openAPI.yaml
openapi: 3.0.3
info:
  title: YoloFarm API - Đồ án Thực tập Đa ngành (CO3109)
  description: |
    Tài liệu đặc tả API cho hệ thống Nông nghiệp thông minh YoloFarm.
    
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Spring Boot Server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/farms:
    get:
      tags: [Farms]
      summary: Lấy danh sách Nông trại
      responses:
        '200':
          description: Lấy danh sách thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FarmResponse'
              examples:
                CoDuLieu:
                  summary: Nông dân có 2 nông trại
                  value:
                    - id: "888e4567-e89b-12d3-a456-426614175555"
                      owner_id: "123e4567-e89b-12d3-a456-426614174000"
                      name: "Vườn Cà Chua Thủy Canh"
                      location: "Khu A"
                      created_at: "2023-10-02T08:30:00Z"
                    - id: "777e4567-e89b-12d3-a456-426614175555"
                      owner_id: "123e4567-e89b-12d3-a456-426614174000"
                      name: "Vườn Dưa Lưới"
                      location: "Khu B"
                      created_at: "2023-10-05T09:00:00Z"
                Trong:
                  summary: Nông dân chưa có nông trại nào
                  value: []
        '401':
          description: Lỗi xác thực
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                ThieuToken:
                  summary: Không gửi header Authorization
                  value:
                    code: 401
                    message: Unauthorized
                    details: "Missing Authorization header."

    post:
      tags: [Farms]
      summary: Tạo Nông trại mới (Dành cho FARMER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarmCreateRequest'
            examples:
              TaoMoi:
                summary: Nhập tên và vị trí vườn
                value:
                  name: "Vườn Cà Chua Thủy Canh"
                  location: "Khu nhà màng A"
      responses:
        '201':
          description: Tạo Nông trại thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmResponse'
              examples:
                ThanhCong:
                  summary: Trả về Object Farm vừa tạo
                  value:
                    id: "888e4567-e89b-12d3-a456-426614175555"
                    owner_id: "123e4567-e89b-12d3-a456-426614174000"
                    name: "Vườn Cà Chua Thủy Canh"
                    location: "Khu nhà màng A"
                    created_at: "2023-10-02T08:30:00Z"
        '400':
          description: Thiếu thông tin bắt buộc
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                LoiThieuTen:
                  summary: Không nhập tên nông trại
                  value:
                    code: 400
                    message: Validation Failed
                    details: "Tên nông trại (name) không được để trống."
        '401':
          description: Chưa mang theo JWT Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                ThieuToken:
                  summary: Không gửi header Authorization
                  value:
                    code: 401
                    message: Unauthorized
                    details: "Missing Authorization header."
...
  /api/v1/farms/{farmId}:
    get:
      tags: [Farms]
      summary: Lấy thông tin chi tiết một Nông trại
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chi tiết Farm
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmResponse'
    put:
      tags: [Farms]
      summary: Cập nhật thông tin Nông trại
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                location:
                  type: string
            examples:
              SuaThongTin:
                summary: Cập nhật tên và vị trí
                value:
                  name: "Vườn Cà Chua (Đã mở rộng)"
                  location: "Khu nhà màng A và B"
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmResponse'
    delete:
      tags: [Farms]
      summary: Xóa một Nông trại
      description: |
        Xóa vĩnh viễn hoặc đánh dấu xóa một nông trại. 
        Lưu ý: Hệ thống sẽ chặn hành động này nếu nông trại vẫn còn thiết bị (Devices) hoặc các kịch bản tự động (Rules) đang gán vào.
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Xóa nông trại thành công (No Content).
        '401':
          description: Chưa đăng nhập.
          content:
            application/json:
              example:
                code: 401
                message: "Unauthorized: Vui lòng đăng nhập để thực hiện."
        '403':
          description: Lỗi phân quyền (Forbidden).
          content:
            application/json:
              example:
                code: 403
                message: "Bạn không phải chủ sở hữu (Owner) nên không có quyền xóa nông trại này."
        '404':
          description: Không tìm thấy (Not Found).
          content:
            application/json:
              example:
                code: 404
                message: "Nông trại không tồn tại hoặc đã bị xóa."
        '409':
          description: Xung đột ràng buộc dữ liệu (Conflict).
          content:
            application/json:
              examples:
                ConThietBi:
                  summary: Lỗi còn thiết bị
                  value:
                    code: 409
                    message: "Không thể xóa: Nông trại hiện vẫn còn 05 thiết bị đang liên kết."
                ConRules:
                  summary: Lỗi còn kịch bản tự động
                  value:
                    code: 409
                    message: "Vui lòng xóa các Rules (Tự động hóa) trước khi xóa Nông trại."
  /api/v1/farms/{farmId}/devices:
    get:
      tags: [Devices]
      summary: (FARMER) Xem danh sách thiết bị trong một nông trại
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Danh sách thiết bị kèm trạng thái kết nối
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceResponse'
              examples:
                DanhSach:
                  summary: Có thiết bị đang mất kết nối
                  value:
                    - id: "d11e4567-e89b-12d3-a456-426614171111"
                      model_id: "m11e4567-e89b-12d3-a456-426614171111"
                      name: "Cảm biến đất khu A"
                      status: "ACTIVE"
                      connection_status: "ONLINE"
                      last_seen: "2023-10-02T10:05:00Z"
                      operating_mode: "AUTO"
                    - id: "d22e4567-e89b-12d3-a456-426614172222"
                      model_id: "m22e4567-e89b-12d3-a456-426614172222"
                      name: "Máy bơm số 1"
                      status: "ACTIVE"
                      connection_status: "OFFLINE"
                      last_seen: "2023-10-01T08:00:00Z"
                      operating_mode: "MANUAL"
...
  /api/v1/farms/{farmId}/rules:
    get:
      tags: [Rules]
      summary: (FARMER) Lấy danh sách các Luật tự động hóa của Nông trại
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Danh sách Rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleResponse'
  /api/v1/rules:
    post:
      tags: [Rules]
      summary: Tạo mới Luật Tự động hóa (Dành cho FARMER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreateRequest'
            examples:
              TuoiTuDong:
                summary: Luật tưới cây khi đất khô
                value:
                  farm_id: "888e4567-e89b-12d3-a456-426614175555"
                  rule_name: "Tưới tự động khi đất khô"
                  trigger_device_id: "cam-bien-do-am-id"
                  operator: "<"
                  threshold_value: 40.5
                  action_device_id: "may-bom-id"
                  action_command: "ON"
      responses:
        '201':
          description: Đã tạo Rule thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
              examples:
                ThanhCong:
                  summary: Object Rule được sinh ra
                  value:
                    id: "r11e4567-e89b-12d3-a456-426614171111"
                    rule_name: "Tưới tự động khi đất khô"
                    is_active: true
        '400':
          description: Lỗi logic khi thiết lập luật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                LoiThietBi:
                  summary: Chọn sai loại thiết bị Action
                  value:
                    code: 400
                    message: Validation Failed
                    details: "Thiết bị thực thi (action_device_id) bắt buộc phải là loại ACTUATOR."
  /api/v1/rules/{ruleId}/toggle:
    patch:
      tags: [Rules]
      summary: Bật/Tắt (Enable/Disable) một Rule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_active:
                  type: boolean
            examples:
              TatLuat:
                summary: Tắt luật tưới cây
                value:
                  is_active: false
      responses:
        '200':
          description: Đã cập nhật trạng thái
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                ThanhCong:
                  summary: Phản hồi thành công
                  value:
                    message: "Đã tắt Rule thành công."
  /api/v1/rules/{ruleId}:
    delete:
      tags: [Rules]
      summary: Xóa một Rule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Xóa thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                XoaThanhCong:
                  summary: Phản hồi xóa
                  value:
                    message: "Đã xóa Rule thành công."

    put:
      tags: [Rules]
      summary: (FARMER) Cập nhật cấu hình Luật Tự động hóa
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreateRequest'
            examples:
              CapNhatNguong:
                summary: Thay đổi ngưỡng độ ẩm từ 40% lên 50%
                value:
                  farm_id: "888e4567-e89b-12d3-a456-426614175555"
                  rule_name: "Tưới tự động khi đất khô (Đã sửa)"
                  trigger_device_id: "cam-bien-do-am-id"
                  operator: "<"
                  threshold_value: 50.0
                  action_device_id: "may-bom-id"
                  action_command: "ON"
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'

  # ==========================================
  # 5. TELEMETRY (DỮ LIỆU VẼ BIỂU ĐỒ)
  # ==========================================
...
  /api/v1/farms/{farmId}/ai-analysis:
    post:
      tags: [AI Analysis]
      summary: Upload ảnh để AI chẩn đoán bệnh lý
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                analysis_type:
                  type: string
                  enum: [LEAF_DISEASE, FRUIT_SORTING]
      responses:
        '200':
          description: Phân tích thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiAnalysisResult'
              examples:
                KetQuaBenhLa:
                  summary: Chẩn đoán bệnh đốm lá
                  value:
                    result_label: "Bệnh đốm lá (Leaf Spot)"
                    confidence_score: 0.95
                    image_url: "https://yolofarm-storage.s3.amazonaws.com/leaf_123.jpg"
        '400':
          description: Lỗi định dạng file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                LoiDinhDang:
                  summary: Upload file PDF thay vì file ảnh
                  value:
                    code: 400
                    message: Bad Request
                    details: "Định dạng file không hỗ trợ. Chỉ chấp nhận .JPG, .PNG."
  /api/v1/farms/{farmId}/ai-logs:
    get:
      tags: [AI Analysis]
      summary: Lấy nhật ký chẩn đoán AI của Nông trại
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Danh sách lịch sử AI
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AiLogResponse'
              examples:
                DanhSachLichSu:
                  summary: Xem lại các lần chụp trước đây
                  value:
                    - id: "a11e4567-e89b-12d3-a456-426614171111"
                      analysis_type: "LEAF_DISEASE"
                      result_label: "Bệnh đốm lá"
                      confidence_score: 0.95
                      image_url: "https://yolofarm.com/images/1.jpg"
                      analyzed_at: "2023-10-01T14:30:00Z"

  # ==========================================
  # 8. NOTIFICATIONS (THÔNG BÁO)
  # ==========================================
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: Lấy danh sách thông báo của User hiện tại
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: "Số thứ tự trang (mặc định: 1)"
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: "Số lượng bản ghi trên mỗi trang (mặc định: 20)"
      responses:
        '200':
          description: Danh sách thông báo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationResponse'
              examples:
                DanhSachThongBao:
                  summary: Nông dân nhận được báo cáo duyệt thiết bị
                  value:
                    - id: "n11e4567-e89b-12d3-a456-426614171111"
                      message: "Thiết bị [Máy bơm nước hồ 1] đã được Admin duyệt và cấp Feed Key."
                      is_read: false
                      created_at: "2023-10-02T10:00:00Z"
  /api/v1/notifications/{notificationId}/read:
    put:
      tags: [Notifications]
      summary: Đánh dấu một thông báo là đã đọc
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                ThanhCong:
                  summary: Phản hồi thành công
                  value:
                    message: "Đã đánh dấu đọc."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        details:
          type: string
...
    FarmCreateRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
    FarmResponse:
      type: object
      required:
        - id
        - owner_id
        - name
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
          description: "ID duy nhất của Nông trại (UUID)"
        owner_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
          description: "ID của người dùng sở hữu nông trại này"
        name:
          type: string
          example: "YoloFarm Chi nhánh Pleiku"
          description: "Tên gọi của nông trại"
        location:
          type: string
          example: "123 Đường Lê Lợi, Pleiku, Gia Lai"
          description: "Địa chỉ hoặc tọa độ địa lý (Có thể để trống)"
        created_at:
          type: string
          format: date-time
          example: "2026-03-21T11:17:15Z"
          description: "Thời điểm nông trại được tạo trên hệ thống"

    # --- SCHEMAS NHÓM THIẾT BỊ (DEVICES) ---
...
    DeviceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        farm_id:
          type: string
          format: uuid
        model_id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [PENDING, ACTIVE, REJECTED]
        adafruit_feed_key:
          type: string
          nullable: true
        connection_status:
          type: string
          enum: [ONLINE, OFFLINE]
          description: "FR3.4: Trạng thái kết nối"
        last_seen:
          type: string
          format: date-time
          nullable: true
        operating_mode:
          type: string
          enum: [AUTO, MANUAL]
          description: "FR4: Chế độ điều khiển"
        is_active:
          type: boolean

    # --- SCHEMAS NHÓM TỰ ĐỘNG HÓA (RULES) ---
    RuleCreateRequest:
      type: object
      required: [farm_id, rule_name, rule_type, action_device_id, action_command]
      properties:
        farm_id:
          type: string
          format: uuid
        rule_name:
          type: string
        rule_type:
          type: string
          enum: [CONDITION, SCHEDULE]
          description: "Loại tự động hóa: Theo cảm biến (CONDITION) hoặc Theo lịch trình (SCHEDULE)"
        trigger_device_id:
          type: string
          format: uuid
          nullable: true
          description: "Chỉ bắt buộc nếu rule_type là CONDITION"
        operator:
          type: string
          enum: ['>', '<', '==']
          nullable: true
        threshold_value:
          type: number
          format: float
          nullable: true
        cron_expression:
          type: string
          nullable: true
          description: "Chỉ bắt buộc nếu rule_type là SCHEDULE (VD: '0 6 * * *' là 6h sáng mỗi ngày)"
        action_device_id:
          type: string
          format: uuid
        action_command:
          type: string
          enum: ['ON', 'OFF']
    RuleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        farm_id:
          type: string
          format: uuid
        rule_name:
          type: string
        rule_type:
          type: string
        trigger_device_id:
          type: string
          format: uuid
          nullable: true
        operator:
          type: string
          nullable: true
        threshold_value:
          type: number
          format: float
          nullable: true
        cron_expression:
          type: string
          nullable: true
        action_device_id:
          type: string
          format: uuid
        action_command:
          type: string
        is_active:
          type: boolean

    # --- SCHEMAS NHÓM TELEMETRY ---
...
    AiAnalysisResult:
      type: object
      properties:
        result_label:
          type: string
          description: "Tên nhãn kết quả phân tích (Ví dụ: Bệnh vàng lá)"
        confidence_score:
          type: number
          format: float
          description: "Độ tự tin của mô hình (từ 0.0 đến 1.0)"
        image_url:
          type: string
          description: "Đường dẫn URL để Frontend load lại ảnh"
    AiLogResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analysis_type:
          type: string
        result_label:
          type: string
        confidence_score:
          type: number
          format: float
        image_url:
          type: string
        analyzed_at:
          type: string
          format: date-time

    # --- SCHEMAS NHÓM NOTIFICATIONS ---
    NotificationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        message:
          type: string
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
...